flashforge_complete_for_ios/requirements.txt
pyqt5==5.15.9
pyserial==3.5
pyusb==1.2.1
adb==1.3.0
fastboot==0.2
requests==2.31.0
tqdm==4.66.1
cryptography==41.0.3